fix: validate object IDs in banner, category, event, and ticket controllers to improve error handling

+++ backend/src/controllers/banner.controller.ts
@@ -2,7 +2,7 @@ import { Response } from "express";
 import { IPaginationQuery, IReqUser } from "../utils/interface";
 import BannerModel, { bannerDAO, TypeBanner } from "../models/banner.model";
 import response from "../utils/response";
-import { FilterQuery } from "mongoose";
+import { FilterQuery, isValidObjectId } from "mongoose";
 
 export default {
   async create(req: IReqUser, res: Response) {
@@ -67,6 +67,10 @@ export default {
     try {
       const { id } = req.params;
 
+      if (!isValidObjectId(id)) {
+        return response.notFound(res, "Banner not found");
+      }
+
       const result = await BannerModel.findById(id);
 
       if (!result) {
@@ -82,6 +86,10 @@ export default {
   async update(req: IReqUser, res: Response) {
     try {
       const { id } = req.params;
+
+      if (!isValidObjectId(id)) {
+        return response.notFound(res, "Banner not found");
+      }
 
       const result = await BannerModel.findByIdAndUpdate(id, req.body, {
         new: true,
@@ -100,6 +108,10 @@ export default {
   async remove(req: IReqUser, res: Response) {
     try {
       const { id } = req.params;
+
+      if (!isValidObjectId(id)) {
+        return response.notFound(res, "Banner not found");
+      }
 
       const result = await BannerModel.findByIdAndDelete(id, {
         new: true,
